requirements.txt
fastapi==0.103.1
uvicorn==0.23.2
python-multipart==0.0.6
google-generativeai==0.3.1
opencv-python==4.8.0.76
pillow==10.0.0
python-dotenv==1.0.0
numpy
pydantic==2.3.0
firebase-admin==6.2.0